requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
cx_Oracle==8.3.0
python-dotenv==1.0.0
google-cloud-bigquery==3.12.0
pymysql==1.1.0
sqlalchemy==2.0.25
psycopg2-binary>=2.9.3